requirements.txt
asgiref==3.9.1
Django==4.2.23
sqlparse==0.5.3
typing_extensions==4.14.1
python-dotenv==1.0.1
djangorestframework==3.15.2
psycopg-binary==3.2.9

djangorestframework-simplejwt>=5.2.2
